+++ src/terralib/maptools/Grouping.h
@@ -131,14 +131,14 @@ namespace te
 
           \output The standard deviation.
         */
-        const float getStdDeviation() const;
+        const double getStdDeviation() const;
 
        /*!
           \brief It sets the standard deviation for the Standard Deviation grouping.
 
           \param stdDeviation The standard deviation.
         */
-        void setStdDeviation(float stdDeviation);
+        void setStdDeviation(double stdDeviation);
 
         /*!
           \brief It gets the vector of grouping itens.
@@ -160,7 +160,7 @@ namespace te
         GroupingType m_type;             //!< The grouping type.
         size_t m_precision;              //!< The precision of the values.
         size_t m_numSlices;              //!< The number of slices used in the Equal Steps and Quantil groupings.
-        float m_stdDeviation;            //!< The standard deviation used in the Standard Deviation grouping.
+        double m_stdDeviation;            //!< The standard deviation used in the Standard Deviation grouping.
 
         std::vector<te::map::GroupingItem*> m_itens; //!< The vector of grouping itens.
     };
